Move DaliOperatorTable to Coral-Common and add abstract class FunctionResolver
